+++ packages/expo-sqlite/android/src/main/java/expo/modules/sqlite/NativeDatabaseBinding.kt
@@ -46,6 +46,7 @@ internal class NativeDatabaseBinding : Closeable {
   external fun sqlite3_changes(): Int
   external fun sqlite3_finalize_all_statement()
   external fun sqlite3_close(): Int
+  external fun sqlite3_interrupt()
   external fun sqlite3_db_filename(databaseName: String): String
   external fun sqlite3_enable_load_extension(onoff: Int): Int
   external fun sqlite3_exec(source: String): Int